feat: Disable scrolling with feature details modal is up
This uses a bit of a hacky workaround to do this.  We set a value
on the $rootScope that the base template can then use to change
a class on the body.

In a future version of angular bootstrap they are supposed to add
back in a way to do this automatically as part of that tool.